fixed stale docs, Windows hash inconsistency, redundant canonicalization, and unsafe gc
- updated spec and Windows doc comment referencing old `.tracey/` paths
- aligned Windows `local_endpoint` to use Blake3 hash from `state_dir()`
  instead of `DefaultHasher` on uncanonicalized path
- extracted `state_dir_inner` to avoid double-canonicalization in
  `ensure_state_dir`
- moved `is_pid_alive` and `read_pid_file_at` to daemon mod.rs so
  `gc` can skip state dirs with a live daemon
- removed stale `-n` shorthand from CLI docs
