+++ docs/content/spec/tracey.md
@@ -928,7 +928,7 @@ The `tracey gc` command MUST remove state directories whose `project-root` metad
 > The `--dry-run` flag MUST cause the command to report what would be removed without deleting anything.
 
 r[daemon.logs.file]
-The daemon MUST write all log output to `.tracey/daemon.log` in the workspace root.
+The daemon MUST write all log output to `daemon.log` in the state directory.
 
 ## Validation
